--- a/src/libraries/Microsoft.Extensions.DependencyInjection.Abstractions/src/ActivatorUtilities.cs
+++ b/src/libraries/Microsoft.Extensions.DependencyInjection.Abstractions/src/ActivatorUtilities.cs
@@ -28,9 +28,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
         // Support caching of constructor metadata for types in collectible assemblies.
         private static readonly Lazy<ConditionalWeakTable<Type, ConstructorInfoEx[]>> s_collectibleConstructorInfos = new();
-#endif
 
-#if NET8_0_OR_GREATER
         // Maximum number of fixed arguments for ConstructorInvoker.Invoke(arg1, etc).
         private const int FixedArgumentThreshold = 4;
 #endif
@@ -66,10 +64,16 @@ namespace Microsoft.Extensions.DependencyInjection
             {
                 constructors = GetOrAddConstructors(instanceType);
             }
+
+            int maxArgs = GetMaxArgCount();
+            object?[] values = new object?[maxArgs * 2];
+            Span<object?> parameterValues = new(values, 0, maxArgs);
 #else
             constructors = CreateConstructorInfoExs(instanceType);
+            object?[]? parameterValues = null;
 #endif
 
+            ConstructorMatcher matcher = default;
             ConstructorInfoEx? constructor;
             IServiceProviderIsService? serviceProviderIsService = provider.GetService<IServiceProviderIsService>();
             // if container supports using IServiceProviderIsService, we try to find the longest ctor that
@@ -79,48 +83,71 @@ namespace Microsoft.Extensions.DependencyInjection
             // instance if all parameters given to CreateInstance only match with a single ctor
             if (serviceProviderIsService != null)
             {
-                int bestLength = -1;
-                bool seenPreferred = false;
-
-                ConstructorMatcher bestMatcher = default;
-                bool multipleBestLengthFound = false;
-
+                // Handle the case where the attribute is used.
                 for (int i = 0; i < constructors.Length; i++)
                 {
-                    constructor = constructors[i];
-                    ConstructorMatcher matcher = new(constructor);
-                    bool isPreferred = constructor.IsPreferred;
-                    int length = matcher.Match(parameters, serviceProviderIsService);
-
-                    if (isPreferred)
+                    if (constructors[i].IsPreferred)
                     {
-                        if (seenPreferred)
+                        for (int j = i + 1; j < constructors.Length; j++)
                         {
-                            ThrowMultipleCtorsMarkedWithAttributeException();
+                            if (constructors[j].IsPreferred)
+                            {
+                                ThrowMultipleCtorsMarkedWithAttributeException();
+                            }
                         }
 
-                        if (length == -1)
+#if !NETCOREAPP
+                        parameterValues = new object?[constructors[i].Parameters.Length];
+#endif
+
+                        matcher = new ConstructorMatcher(constructors[i], parameterValues);
+                        if (matcher.Match(parameters, serviceProviderIsService) == -1)
                         {
                             ThrowMarkedCtorDoesNotTakeAllProvidedArguments();
                         }
 
-                        seenPreferred = true;
+                        return matcher.CreateInstance(provider);
+                    }
+                }
+
+                int bestLength = -1;
+                ConstructorMatcher bestMatcher = default;
+                bool multipleBestLengthFound = false;
+#if NETCOREAPP
+                Span<object?> bestParameterValues = new(values, maxArgs, maxArgs);
+#endif
+
+                // Find the constructor with the most matches.
+                for (int i = 0; i < constructors.Length; i++)
+                {
+                    constructor = constructors[i];
+
+#if NETCOREAPP
+                    parameterValues.Clear();
+#else
+                    parameterValues = new object?[constructor.Parameters.Length];
+#endif
+
+                    matcher = new ConstructorMatcher(constructor, parameterValues);
+                    int length = matcher.Match(parameters, serviceProviderIsService);
+
+                    Debug.Assert(!constructor.IsPreferred);
+
+                    if (bestLength < length)
+                    {
                         bestLength = length;
-                        bestMatcher = matcher;
+#if NETCOREAPP
+                        parameterValues.CopyTo(bestParameterValues);
+#else
+                        object?[]? bestParameterValues = new object?[length];
+                        parameterValues.CopyTo(bestParameterValues, index: 0);
+#endif
+                        bestMatcher = new ConstructorMatcher(matcher.ConstructorInfo, bestParameterValues);
                         multipleBestLengthFound = false;
                     }
-                    else if (!seenPreferred)
+                    else if (bestLength == length)
                     {
-                        if (bestLength < length)
-                        {
-                            bestLength = length;
-                            bestMatcher = matcher;
-                            multipleBestLengthFound = false;
-                        }
-                        else if (bestLength == length)
-                        {
-                            multipleBestLengthFound = true;
-                        }
+                        multipleBestLengthFound = true;
                     }
                 }
 
@@ -149,24 +176,25 @@ namespace Microsoft.Extensions.DependencyInjection
                 }
             }
 
-            FindApplicableConstructor(instanceType, argumentTypes, out ConstructorInfo constructorInfo, out int?[] parameterMap);
+            FindApplicableConstructor(instanceType, argumentTypes, constructors, out ConstructorInfo constructorInfo, out int?[] parameterMap);
+            constructor = FindConstructorEx(constructorInfo, constructors);
 
-            // Find the ConstructorInfoEx from the given constructorInfo.
-            constructor = null;
-            foreach (ConstructorInfoEx ctor in constructors)
+            matcher = new ConstructorMatcher(constructor, parameterValues);
+            matcher.MapParameters(parameterMap, parameters);
+            return matcher.CreateInstance(provider);
+
+#if NETCOREAPP
+            int GetMaxArgCount()
             {
-                if (ReferenceEquals(ctor.Info, constructorInfo))
+                int max = 0;
+                for (int i = 0; i < constructors.Length; i++)
                 {
-                    constructor = ctor;
-                    break;
+                    max = int.Max(max, constructors[i].Parameters.Length);
                 }
+
+                return max;
             }
-
-            Debug.Assert(constructor != null);
-
-            var constructorMatcher = new ConstructorMatcher(constructor);
-            constructorMatcher.MapParameters(parameterMap, parameters);
-            return constructorMatcher.CreateInstance(provider);
+#endif
         }
 
 #if NETCOREAPP
@@ -280,7 +308,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
         private static void CreateFactoryInternal([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type instanceType, Type[] argumentTypes, out ParameterExpression provider, out ParameterExpression argumentArray, out Expression factoryExpressionBody)
         {
-            FindApplicableConstructor(instanceType, argumentTypes, out ConstructorInfo constructor, out int?[] parameterMap);
+            FindApplicableConstructor(instanceType, argumentTypes, constructors: null, out ConstructorInfo constructor, out int?[] parameterMap);
 
             provider = Expression.Parameter(typeof(IServiceProvider), "provider");
             argumentArray = Expression.Parameter(typeof(object[]), "argumentArray");
@@ -401,10 +429,10 @@ namespace Microsoft.Extensions.DependencyInjection
             [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type instanceType,
             Type?[] argumentTypes)
         {
-            FindApplicableConstructor(instanceType, argumentTypes, out ConstructorInfo constructor, out int?[] parameterMap);
+            FindApplicableConstructor(instanceType, argumentTypes, constructors: null, out ConstructorInfo constructor, out int?[] parameterMap);
             Type declaringType = constructor.DeclaringType!;
 
-#if NET8_0_OR_GREATER
+#if NETCOREAPP
             ConstructorInvoker invoker = ConstructorInvoker.Create(constructor);
 
             ParameterInfo[] constructorParameters = constructor.GetParameters();
@@ -473,7 +501,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
             FactoryParameterContext[] parameters = GetFactoryParameterContext();
             return (serviceProvider, arguments) => ReflectionFactoryCanonical(constructor, parameters, declaringType, serviceProvider, arguments);
-#endif // NET8_0_OR_GREATER
+#endif // NETCOREAPP
 
             FactoryParameterContext[] GetFactoryParameterContext()
             {
@@ -518,13 +546,14 @@ namespace Microsoft.Extensions.DependencyInjection
         private static void FindApplicableConstructor(
             [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type instanceType,
             Type?[] argumentTypes,
+            ConstructorInfoEx[]? constructors,
             out ConstructorInfo matchingConstructor,
             out int?[] matchingParameterMap)
         {
             ConstructorInfo? constructorInfo;
             int?[]? parameterMap;
 
-            if (!TryFindPreferredConstructor(instanceType, argumentTypes, out constructorInfo, out parameterMap) &&
+            if (!TryFindPreferredConstructor(instanceType, argumentTypes, constructors, out constructorInfo, out parameterMap) &&
                 !TryFindMatchingConstructor(instanceType, argumentTypes, out constructorInfo, out parameterMap))
             {
                 throw new InvalidOperationException(SR.Format(SR.CtorNotLocated, instanceType));
@@ -532,6 +561,21 @@ namespace Microsoft.Extensions.DependencyInjection
 
             matchingConstructor = constructorInfo;
             matchingParameterMap = parameterMap;
+        }
+
+        // Find the ConstructorInfoEx from the given constructorInfo.
+        private static ConstructorInfoEx FindConstructorEx(ConstructorInfo constructorInfo, ConstructorInfoEx[] constructorExs)
+        {
+            for (int i = 0; i < constructorExs.Length; i++)
+            {
+                if (ReferenceEquals(constructorExs[i].Info, constructorInfo))
+                {
+                    return constructorExs[i];
+                }
+            }
+
+            Debug.Assert(false);
+            return null!;
         }
 
         // Tries to find constructor based on provided argument types
@@ -571,6 +615,7 @@ namespace Microsoft.Extensions.DependencyInjection
         private static bool TryFindPreferredConstructor(
             [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type instanceType,
             Type?[] argumentTypes,
+            ConstructorInfoEx[]? constructors,
             [NotNullWhen(true)] out ConstructorInfo? matchingConstructor,
             [NotNullWhen(true)] out int?[]? parameterMap)
         {
@@ -578,21 +623,33 @@ namespace Microsoft.Extensions.DependencyInjection
             matchingConstructor = null;
             parameterMap = null;
 
-            foreach (ConstructorInfo? constructor in instanceType.GetConstructors())
+            if (constructors is null)
             {
-                if (constructor.IsDefined(typeof(ActivatorUtilitiesConstructorAttribute), false))
+#if NETCOREAPP
+                if (!s_constructorInfos.TryGetValue(instanceType, out constructors))
+                {
+                    constructors = GetOrAddConstructors(instanceType);
+                }
+#else
+                constructors = CreateConstructorInfoExs(instanceType);
+#endif
+            }
+
+            foreach (ConstructorInfoEx constructor in constructors)
+            {
+                if (constructor.IsPreferred)
                 {
                     if (seenPreferred)
                     {
                         ThrowMultipleCtorsMarkedWithAttributeException();
                     }
 
-                    if (!TryCreateParameterMap(constructor.GetParameters(), argumentTypes, out int?[] tempParameterMap))
+                    if (!TryCreateParameterMap(constructor.Info.GetParameters(), argumentTypes, out int?[] tempParameterMap))
                     {
                         ThrowMarkedCtorDoesNotTakeAllProvidedArguments();
                     }
 
-                    matchingConstructor = constructor;
+                    matchingConstructor = constructor.Info;
                     parameterMap = tempParameterMap;
                     seenPreferred = true;
                 }
@@ -649,6 +706,17 @@ namespace Microsoft.Extensions.DependencyInjection
             public readonly ParameterInfo[] Parameters;
             public readonly bool IsPreferred;
             private readonly object?[]? _parameterKeys;
+#if NETCOREAPP
+            public ConstructorInvoker? _invoker;
+            public ConstructorInvoker Invoker
+            {
+                get
+                {
+                    _invoker ??= ConstructorInvoker.Create(Info);
+                    return _invoker;
+                }
+            }
+#endif
 
             public ConstructorInfoEx(ConstructorInfo constructor)
             {
@@ -710,16 +778,23 @@ namespace Microsoft.Extensions.DependencyInjection
             }
         }
 
-        private readonly struct ConstructorMatcher
+        private readonly ref struct ConstructorMatcher
         {
             private readonly ConstructorInfoEx _constructor;
-            private readonly object?[] _parameterValues;
 
-            public ConstructorMatcher(ConstructorInfoEx constructor)
+#if NETCOREAPP
+            private readonly Span<object?> _parameterValues;
+            public ConstructorMatcher(ConstructorInfoEx constructor, Span<object?> parameterValues)
+#else
+            private readonly object?[] _parameterValues;
+            public ConstructorMatcher(ConstructorInfoEx constructor, object?[] parameterValues)
+#endif
             {
                 _constructor = constructor;
-                _parameterValues = new object[constructor.Parameters.Length];
+                _parameterValues = parameterValues;
             }
+
+            public ConstructorInfoEx ConstructorInfo => _constructor;
 
             public int Match(object[] givenParameters, IServiceProviderIsService serviceProviderIsService)
             {
@@ -790,7 +865,9 @@ namespace Microsoft.Extensions.DependencyInjection
                     }
                 }
 
-#if NETFRAMEWORK || NETSTANDARD2_0
+#if NETCOREAPP
+                return _constructor.Invoker.Invoke(_parameterValues.Slice(0, _constructor.Parameters.Length));
+#else
                 try
                 {
                     return _constructor.Info.Invoke(_parameterValues);
@@ -801,8 +878,6 @@ namespace Microsoft.Extensions.DependencyInjection
                     // The above line will always throw, but the compiler requires we throw explicitly.
                     throw;
                 }
-#else
-                return _constructor.Info.Invoke(BindingFlags.DoNotWrapExceptions, binder: null, parameters: _parameterValues, culture: null);
 #endif
             }
 
@@ -828,7 +903,7 @@ namespace Microsoft.Extensions.DependencyInjection
             throw new InvalidOperationException(SR.Format(SR.MarkedCtorMissingArgumentTypes, nameof(ActivatorUtilitiesConstructorAttribute)));
         }
 
-#if NET8_0_OR_GREATER // Use the faster ConstructorInvoker which also has alloc-free APIs when <= 4 parameters.
+#if NETCOREAPP // Use the faster ConstructorInvoker which also has alloc-free APIs when <= 4 parameters.
         private static object ReflectionFactoryServiceOnlyFixed(
             ConstructorInvoker invoker,
             FactoryParameterContext[] parameters,
@@ -1101,7 +1176,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
             return constructor.Invoke(BindingFlags.DoNotWrapExceptions, binder: null, constructorArguments, culture: null);
         }
-#endif // NET8_0_OR_GREATER
+#endif
 
 #if NETCOREAPP
         internal static class ActivatorUtilitiesUpdateHandler