@stephentoub 2024-03-07
> verifying tests

Given the nature of the change, it'd be good to see perf test results on both .NET 9 and .NET Framework.